Update docs and examples for new styleguide, reorganize some files for new file structure
