+++ src/components/forms/cdrInput.md
@@ -1,0 +1,10 @@
+**NOTE:** `v-model` is required.
+
+## Events
+| Name | Type | Description | |
+| --- | --- | --- |
+| `input` | String | Current input value. Fires while typing. |
+| `change` | String | New input value (if changed). Fires on blur. |
+| `blur` | Event | Fires when input loses focus. |
+| `focus` | Event | Fires when input gains focus. |
+| `paste` | Event | Fires when text is pasted into input. |
