+++ src/components/icons/cdrIcon-hover.backstop.js
