+++ src/components/icons/cdrIcon.backstop.js
