+++ src/examples/buttons/demo/Black.vue
@@ -3,6 +3,7 @@
     <cdr-button data-backstop="cdr-button--alternate" modifier="alternate">Black alternate</cdr-button>
     <cdr-button modifier="alternate small">Black alternate small</cdr-button>
     <cdr-button modifier="alternate extra-small">Black alternate extra small</cdr-button>
+    <cdr-button modifier="alternate" disabled>Black Disabled</cdr-button>
   </cdr-button-group>
 </template>
