+++ src/examples/buttons/demo/Default.vue
@@ -3,6 +3,7 @@
     <cdr-button data-backstop="cdr-button" type="submit" :on-click="log">Default with click</cdr-button>
     <cdr-button type="submit" modifier="small">Default small submit</cdr-button>
     <cdr-button type="reset" modifier="extra-small">Default extra small reset</cdr-button>
+    <cdr-button disabled>Default Disabled</cdr-button>
   </cdr-button-group>
 </template>
