+++ src/examples/buttons/demo/Secondary.vue
@@ -3,6 +3,7 @@
     <cdr-button data-backstop="cdr-button--secondary" modifier="secondary">Secondary</cdr-button>
     <cdr-button modifier="secondary small">Secondary small</cdr-button>
     <cdr-button modifier="secondary extra-small">Secondary extra small</cdr-button>
+    <cdr-button modifier="secondary" disabled>Secondary Disabled</cdr-button>
   </cdr-button-group>
 </template>
