--- src/examples/forms/form.play.js
@@ -1,14 +1,0 @@
-import { play } from 'vue-play';
-import Form from './Form';
-import FormExample from '!raw-loader!./Form';
-
-import TextDoc from './Input.md';
-
-play(Form)
-  .displayName('Forms')
-  .add('Input', {
-    ...Form,
-    example: FormExample,
-    readme: TextDoc,
-  })
-  ;
