--- src/examples/lists/List.md
@@ -1,6 +1,0 @@
-## API
-
-| Prop | Type | Default | Required | Values | Description |
-| --- | --- | --- | --- | --- | --- |
-| `:type` | String | 'ul' |  | {'ul', 'ol'} | Defines the list type, unordered/ordered |
-| `:modifier` | Array (of Strings) |  |  | {'unstyled', 'bulleted', 'numbered', 'menu'} | Style modifiers for lists |
